remove node api
